Урок №7

+++ src/news/comments/comments.service.ts
@@ -5,6 +5,11 @@ import { CommentsEntity } from './comments.entity';
 import { Repository } from 'typeorm';
 import { NewsService } from '../news.service';
 import { UsersService } from '../../users/users.service';
+import { EventsComment } from './EventsComment.enum';
+import {
+  checkPermission,
+  Modules,
+} from '../../auth/role/utils/check-permission';
 
 export type Comment = {
   id?: number;
@@ -27,7 +32,6 @@ export class CommentsService {
     private readonly userService: UsersService,
     private readonly eventEmitter: EventEmitter2,
   ) {}
-  private readonly comments = {};
 
   async create(
     idNews: number,
@@ -64,9 +68,18 @@ export class CommentsService {
   }
 
   async edit(idComment: number, comment: CommentEdit): Promise<CommentsEntity> {
-    const _comment = await this.commentsRepository.findOne(idComment);
+    const _comment = await this.commentsRepository.findOne({
+      where: { id: idComment },
+      relations: ['news', 'user'],
+    });
     _comment.message = comment.message;
-    return this.commentsRepository.save(_comment);
+    const _updatedComment = await this.commentsRepository.save(_comment);
+    this.eventEmitter.emit(EventsComment.edit, {
+      commentId: idComment,
+      newsId: _comment.news.id,
+      comment: _updatedComment,
+    });
+    return _updatedComment;
   }
 
   async findAll(idNews: number): Promise<CommentsEntity[]> {
@@ -76,10 +89,10 @@ export class CommentsService {
     });
   }
 
-  async remove(idComment: number): Promise<CommentsEntity> {
+  async remove(idComment: number, userId: number): Promise<CommentsEntity> {
     const _comment = await this.commentsRepository.findOne({
       where: { id: idComment },
-      relations: ['news'],
+      relations: ['news', 'user'],
     });
     if (!_comment) {
       throw new HttpException(
@@ -91,8 +104,24 @@ export class CommentsService {
       );
     }
 
+    const _user = await this.userService.findById(userId);
+    console.log(_user);
+    console.log(_comment.user);
+    if (
+      _user.id !== _comment.user.id &&
+      !checkPermission(Modules.editComment, _user.roles)
+    ) {
+      throw new HttpException(
+        {
+          status: HttpStatus.FORBIDDEN,
+          error: 'Недостаточно прав для удаления',
+        },
+        HttpStatus.FORBIDDEN,
+      );
+    }
+
     const comment = await this.commentsRepository.remove(_comment);
-    this.eventEmitter.emit('comment.remove', {
+    this.eventEmitter.emit(EventsComment.remove, {
       commentId: idComment,
       newsId: _comment.news.id,
     });
